fix: plan mode code review issues
- Add Feedback field to Plan proto message (field 7)
- Reject(): add state guard (approved/executing/completed/rejected → error), store feedback
- Move ApprovePlan/RejectPlan RPC handlers from plans.go to service.go
- Update tests: verify feedback stored, verify state guard for reject
